Add tests for review step
