+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_step/es_stats/es_stats_error.tsx
@@ -69,7 +69,7 @@ export const EsStatsErrors: React.FunctionComponent<Props> = ({ error }) => {
           size="m"
           content={message}
           iconProps={{
-            'data-test-subj': 'requestErrorIconTip',
+            'data-test-subj': 'esRequestErrorIconTip',
           }}
         />
       );
